+++ src/test/java/io/github/astrapi69/swing/menu/factory/JMenuBarFactoryWithNestedMenuTest.java
@@ -32,7 +32,9 @@ import java.io.File;
 import java.util.LinkedHashMap;
 import java.util.Map;
 
-import javax.swing.*;
+import javax.swing.JFrame;
+import javax.swing.JMenuBar;
+import javax.swing.KeyStroke;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -51,6 +53,7 @@ import io.github.astrapi69.swing.menu.enumeration.MenuType;
 import io.github.astrapi69.swing.menu.model.KeyStrokeInfo;
 import io.github.astrapi69.swing.menu.model.MenuInfo;
 import io.github.astrapi69.swing.menu.model.transform.MenuInfoTreeNodeConverter;
+import io.github.astrapi69.swing.menu.model.transform.MenuItemInfoConverter;
 import io.github.astrapi69.throwable.RuntimeExceptionDecorator;
 import io.github.astrapi69.window.adapter.CloseWindow;
 
@@ -149,7 +152,7 @@ public class JMenuBarFactoryWithNestedMenuTest
 
 		idGenerator = LongIdGenerator.of(0L);
 
-		menuBarInfo = MenuInfoTreeNodeConverter.fromJMenuBar();
+		menuBarInfo = MenuItemInfoConverter.fromJMenuBar();
 
 		menuBarTreeNode = BaseTreeNode.<MenuInfo, Long> builder().id(idGenerator.getNextId())
 			.value(menuBarInfo).build();